b10705052/requirements.txt
openai-agents
openai-agents[litellm]
pydantic
python-dotenv